[gocat] optimize code structure
