+++ lib/go/ccat/darwin/libcatclientstatic.a
