+++ lib/go/ccat/include/cat/client_common.h
