+++ lib/go/ccat/include/cat/metric.h
