+++ lib/go/ccat/include/cat/metric_helper.h
