+++ lib/go/ccat/include/cat/trace.h
