+++ lib/go/ccat/linux/libcatclientstatic.a
