+++ lib/go/ccat/transaction.go
@@ -16,28 +16,42 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package gocat
+package ccat
 
 import (
-	"ccat"
+	"time"
 )
 
-type Config struct {
-	host       string
-	port       string
-	domain     string
-	routerType int
+type Transaction struct {
+	Message
+
+	durationInNano      int64
+	durationStartInNano int64
 }
 
-func Init(domain string) {
-	ccat.Init(domain)
-	go ccat.Background()
+func NewTransaction(mtype, name string, flush Flush) *Transaction {
+	return &Transaction{
+		Message:             *NewMessage(mtype, name, flush),
+		durationStartInNano: time.Now().UnixNano(),
+	}
 }
 
-func Shutdown() {
-	ccat.Shutdown()
+func (t *Transaction) Complete() {
+	if t.durationInNano == 0 {
+		durationNano := time.Now().UnixNano() - t.durationStartInNano
+		t.durationInNano = durationNano
+	}
+	t.Message.flush(t)
 }
 
-func Wait() {
-	ccat.Wait()
+func (t *Transaction) GetDuration() int64 {
+	return t.durationInNano
+}
+
+func (t *Transaction) SetDuration(durationInNano int64) {
+	t.durationInNano = durationInNano
+}
+
+func (t *Transaction) SetDurationStart(durationStartInNano int64) {
+	t.durationStartInNano = durationStartInNano
 }
