+++ lib/go/gocat/api.go
@@ -21,8 +21,7 @@ package gocat
 import (
 	"time"
 
-	"ccat"
-	"message"
+	"github.com/dianping/cat/lib/go/ccat"
 )
 
 type catInstance struct {
@@ -32,33 +31,33 @@ func Instance() *catInstance {
 	return &catInstance{}
 }
 
-func (t *catInstance) flush(m message.Messager) {
+func (t *catInstance) flush(m ccat.Messager) {
 	ccat.Send(m)
 }
 
-func (t *catInstance) NewTransaction(mtype, name string) *message.Transaction {
-	return message.NewTransaction(mtype, name, t.flush)
+func (t *catInstance) NewTransaction(mtype, name string) *ccat.Transaction {
+	return ccat.NewTransaction(mtype, name, t.flush)
 }
 
 func (t *catInstance) NewCompletedTransactionWithDuration(mtype, name string, durationInNano int64) {
 	var trans = t.NewTransaction(mtype, name)
 	trans.SetDuration(durationInNano)
-	if durationInNano > 0 && durationInNano < 60 * time.Second.Nanoseconds() {
+	if durationInNano > 0 && durationInNano < 60*time.Second.Nanoseconds() {
 		trans.SetTimestamp(time.Now().UnixNano() - durationInNano)
 	}
-	trans.SetStatus(message.SUCCESS)
+	trans.SetStatus(ccat.SUCCESS)
 	trans.Complete()
 }
 
-func (t *catInstance) NewEvent(mtype, name string) *message.Event {
-	return &message.Event{
-		Message: *message.NewMessage(mtype, name, t.flush),
+func (t *catInstance) NewEvent(mtype, name string) *ccat.Event {
+	return &ccat.Event{
+		Message: *ccat.NewMessage(mtype, name, t.flush),
 	}
 }
 
-func (t *catInstance) NewHeartbeat(mtype, name string) *message.Heartbeat {
-	return &message.Heartbeat{
-		Message: *message.NewMessage(mtype, name, t.flush),
+func (t *catInstance) NewHeartbeat(mtype, name string) *ccat.Heartbeat {
+	return &ccat.Heartbeat{
+		Message: *ccat.NewMessage(mtype, name, t.flush),
 	}
 }
 
@@ -80,7 +79,7 @@ func (t *catInstance) LogError(err error, args ...string) {
 	}
 	var e = t.NewEvent("Exception", category)
 	var buf = newStacktrace(2, err)
-	e.SetStatus(message.FAIL)
+	e.SetStatus(ccat.FAIL)
 	e.AddData(buf.String())
 	e.Complete()
 }
