+++ lib/go/gocat/api_test.go
@@ -1,7 +1,7 @@
 package gocat
 
 import (
-	"message"
+	"github.com/dianping/cat/lib/go/ccat"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -13,10 +13,10 @@ func Test_NewMessage(t *testing.T) {
 	trans := tree.NewTransaction("foo", "bar")
 	assert.Equal(t, trans.Type, "foo")
 	assert.Equal(t, trans.Name, "bar")
-	assert.Equal(t, trans.Status, message.SUCCESS)
+	assert.Equal(t, trans.Status, ccat.SUCCESS)
 
 	event := tree.NewEvent("foo", "bar")
 	assert.Equal(t, event.Type, "foo")
 	assert.Equal(t, event.Name, "bar")
-	assert.Equal(t, event.Status, message.SUCCESS)
+	assert.Equal(t, event.Status, ccat.SUCCESS)
 }
