+++ lib/go/gocat/const.go
@@ -1,0 +1,9 @@
+package gocat
+
+import "github.com/dianping/cat/lib/go/ccat"
+
+const (
+	SUCCESS = ccat.SUCCESS
+	FAIL    = ccat.FAIL
+	VERSION = "3.0.0"
+)
