+++ lib/go/gocat/trace.go
