+++ lib/go/gocat/trace_test.go
