--- lib/go/src/gocat/const.go
@@ -1,9 +1,0 @@
-package gocat
-
-import "message"
-
-const (
-	SUCCESS = message.SUCCESS
-	FAIL = message.FAIL
-	VERSION = "3.0.0"
-)
