+++ lib/go/test/cat_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 	"time"
 
-	"gocat"
+	"github.com/dianping/cat/lib/go/gocat"
 )
 
 var cat = gocat.Instance()
@@ -23,15 +23,15 @@ func case1() {
 	t.AddData("testcase")
 	t.AddData("foo", "bar")
 	t.SetStatus(gocat.FAIL)
-	t.SetDurationStart(time.Now().UnixNano() - time.Second.Nanoseconds() * 5)
+	t.SetDurationStart(time.Now().UnixNano() - time.Second.Nanoseconds()*5)
 	t.SetTimestamp(time.Now().UnixNano() - time.Second.Nanoseconds())
 	t.SetDuration(time.Millisecond.Nanoseconds() * 1000)
 }
 
 // send completed transaction with duration
 func case2() {
-	cat.NewCompletedTransactionWithDuration(TTYPE, "completed", time.Second.Nanoseconds() * 24)
-	cat.NewCompletedTransactionWithDuration(TTYPE, "completed-over-60s", time.Second.Nanoseconds() * 65)
+	cat.NewCompletedTransactionWithDuration(TTYPE, "completed", time.Second.Nanoseconds()*24)
+	cat.NewCompletedTransactionWithDuration(TTYPE, "completed-over-60s", time.Second.Nanoseconds()*65)
 }
 
 // send event
@@ -55,7 +55,7 @@ func case4() {
 func case5() {
 	cat.LogMetricForCount("metric-1")
 	cat.LogMetricForCount("metric-2", 3)
-	cat.LogMetricForDuration("metric-3", 150 * time.Millisecond.Nanoseconds())
+	cat.LogMetricForDuration("metric-3", 150*time.Millisecond.Nanoseconds())
 }
 
 func run(f func()) {
